+++ src/dbcppp/Signal.h
@@ -57,8 +57,8 @@ namespace dbcppp
 		virtual std::string getUnit() const = 0;
 		virtual bool hasReceiver(const std::string& name) const = 0;
 		virtual std::vector<const std::string*> getReceivers() const = 0;
-		virtual const std::string* getValueDescriptionById(uint64_t id) const = 0;
-		virtual std::vector<std::pair<uint64_t, const std::string*>> getValueDescriptions() const = 0;
+		virtual const std::string* getValueDescriptionById(double id) const = 0;
+		virtual std::vector<std::pair<double, const std::string*>> getValueDescriptions() const = 0;
 		virtual const Attribute* getAttributeValueByName(const std::string& name) const = 0;
 		virtual std::vector<std::pair<std::string, const Attribute*>> getAttributeValues() const = 0;
 		virtual const std::string& getComment() const = 0;
